Fix compiler warnings.

+++ source/external/libdeflate/lib/decompress_template.h
@@ -294,7 +294,7 @@ have_decode_tables:
 		 * end-of-block length, so subtract 1 and it turn it into
 		 * SIZE_MAX.  */
 		STATIC_ASSERT(HUFFDEC_END_OF_BLOCK_LENGTH == 0);
-		if (unlikely((size_t)length - 1 >= out_end - out_next)) {
+		if (unlikely((size_t)length - 1 >= (size_t)(out_end - out_next))) {
 			if (unlikely(length != HUFFDEC_END_OF_BLOCK_LENGTH))
 				return LIBDEFLATE_INSUFFICIENT_SPACE;
 			goto block_done;
@@ -328,7 +328,7 @@ have_decode_tables:
 
 		/* The match source must not begin before the beginning of the
 		 * output buffer.  */
-		SAFETY_CHECK(offset <= out_next - (const u8 *)out);
+		SAFETY_CHECK(offset <= (size_t)(out_next - (const u8 *)out));
 
 		/*
 		 * Copy the match: 'length' bytes at 'out_next - offset' to
